requirements.txt
streamlit
moviepy
openai
python-dotenv
pillow
pydub
pygame
phi
